feat: 174580981 rename additionalLocales to SEOlocales
