+++ packages/mwp-config/index.js
@@ -1,7 +1,7 @@
 module.exports = {
 	env: require('./env'),
 	locales: require('./locales'),
-	additionalLocales: require('./additionalLocales'),
+	seoLocales: require('./seoLocales'),
 	localesSecondary: require('./localesSecondary'),
 	package: require('./package'),
 	paths: require('./paths'),
