+++ packages/mwp-config/seoLocales.js
